requirements.txt
asgiref==3.5.2
autopep8==2.0.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-cors-headers==3.7.0
django-phonenumber-field==7.0.0
django-redis-cache==3.0.0
django-redis-sessions==0.6.2
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
openapi-codec==1.3.2
phonenumbers==8.13.0
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
redis==3.5.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
